@@ -16,14 +16,14 @@
 	</modules>
  
  	<properties>
-		<mybatis.jpa.extra.version>2.8</mybatis.jpa.extra.version><!--project version-->
+		<mybatis.jpa.extra.version>3.0</mybatis.jpa.extra.version><!--project version-->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<jdk.version>1.8</jdk.version>
-		<servlet-api.version>3.1.0</servlet-api.version>
-		<junit.version>4.13.1</junit.version>
+		<jdk.version>17</jdk.version>
+		<servlet-api.version>6.0.0</servlet-api.version>
+		<junit.version>5.9.2</junit.version>
 		
-		<spring.version>5.3.22</spring.version>
-		<springboot.version>2.6.10</springboot.version>
+		<spring.version>6.0.4</spring.version>
+		<springboot.version>3.0.1</springboot.version>
 		
 		<commons-collections.version>3.2.2</commons-collections.version>
 		<beanutils.version>1.9.4</beanutils.version>
