@@ -22,17 +22,14 @@ import java.util.List;
 import org.apache.mybatis.jpa.test.dao.service.StudentsService;
 import org.apache.mybatis.jpa.test.entity.Students;
 import org.apache.mybatis.jpa.util.JpaWebContext;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationContext;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest(classes = MybatisJpaApplication.class)
 public class MybatisJpaApplicationTest{ 
     private static final Logger _logger = LoggerFactory.getLogger(MybatisJpaApplicationTest.class);
@@ -47,10 +44,10 @@ public class MybatisJpaApplicationTest{
     private ApplicationContext applicationContext;
  
     
-    @Before
+    @BeforeEach
     public  void before() {
     	_logger.info("---------------- before");
-    	JpaWebContext.applicationContext=applicationContext;
+    	JpaWebContext.applicationContext = applicationContext;
 
     }
     
