@@ -22,8 +22,8 @@ import java.util.List;
 
 import org.apache.mybatis.jpa.test.dao.service.StudentsService;
 import org.apache.mybatis.jpa.test.entity.Students;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -137,8 +137,8 @@ public class CurdTestRunner {
 		 }
 	}
 	
-	@Before
-	public void initSpringContext(){
+	@BeforeAll
+	public static void initSpringContext(){
 		if(InitContext.context!=null) return;
 		service = new InitContext().init();
 	}
