@@ -23,8 +23,8 @@ import java.util.List;
 import org.apache.mybatis.jpa.query.Query;
 import org.apache.mybatis.jpa.test.dao.service.StudentsService;
 import org.apache.mybatis.jpa.test.entity.Students;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -69,8 +69,8 @@ public class QueryTestRunner {
 		 }
 	}
 	
-	@Before
-	public void initSpringContext(){
+	@BeforeAll
+	public static void initSpringContext(){
 		if(InitContext.context!=null) return;
 		service = new InitContext().init();
 	}
