@@ -19,7 +19,7 @@ package org.apache.mybatis.jpa.test;
 
 import org.apache.mybatis.jpa.id.SerialGenerator;
 import org.apache.mybatis.jpa.util.MacAddress;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
