@@ -18,7 +18,7 @@
 package org.apache.mybatis.jpa.test;
 
 import org.apache.mybatis.jpa.id.SnowFlakeIdGenerator;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
